refactor: 컴포넌트 분류, 변수명 변경, media-query추가
